+++ CHANGELOG.md
@@ -14,6 +14,9 @@ All notable changes to this project will be documented in this file. Dates are d
   - Fix error of number input for ios keyboard
 - Trade Screen
   - Swap currencies button has been added
+- Redux Store
+  - Flipper debugger included
+  - Debugger dependency has been removed from prod build
 
 > 4 Dec 2022
 - New signIn and signUp flows